feat: macOS (MPS) support and bug fixes
- Add backend/app/utils/device.py for platform-aware device/dtype handling
- Remove bitsandbytes from requirements.txt; add requirements-cuda.txt for CUDA envs
- Fix bfloat16 → NumPy conversion (.float() before .numpy())
- Fix extraction error details propagating to frontend (error field in JobStatusResponse)
- Add polling timeout (10min) and network error detection in frontend
- Add layer parameter range validation (ge=-100, le=100)
- Fix docs: remove server-specific hardware references, correct FAQ, add admin setup section

+++ backend/app/services/extraction.py
@@ -12,7 +12,7 @@ def extract_hidden_states_from_generation(outputs, input_length: int, layer_idx:
     for token_idx in range(len(outputs.hidden_states)):
         step_hidden = outputs.hidden_states[token_idx]
         layer_hidden = step_hidden[layer_idx]
-        token_hidden = layer_hidden[0, -1, :].cpu().numpy()
+        token_hidden = layer_hidden[0, -1, :].cpu().float().numpy()
         hidden_states_list.append(token_hidden)
     return np.stack(hidden_states_list, axis=0)
